Rename src/physics.ts -> src/player-state.ts to disambiguate
There were two files named physics.ts at different levels: the player-state
integration layer (src/physics.ts) and the physics math kernel
(src/snowman/physics.ts). The top-level file holds no physics math — it owns the
typed PlayerState container and the step/reset wiring over the kernel — so the
shared name was the source of confusion.

Rename it to player-state.ts and the matching test
(tests/physics-state-tests.js -> tests/player-state-tests.js, plus its
test:physics script). Update all import specifiers, the contract-surface Section E
seam, the coverage-merge LCOV fixtures, and stale doc-comments. The snowman/physics.ts
kernel and the exported `Physics` API object are intentionally left unchanged.
